requirements.txt
absl-py==2.1.0
attrs==25.3.0
bleak==0.22.3
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
exif==1.6.1
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frame-sdk==1.2.4
frameutils==0.1.15
fsspec==2025.2.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
plum-py==0.8.7
protobuf==4.25.6
pycparser==2.22
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.0
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.8.0
simpleaudio==1.0.4
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
